+++ my-financial-vite/package.json
@@ -21,6 +21,7 @@
     "postcss": "^8.5.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.2.0",
     "recharts": "^2.15.1",
     "tailwindcss": "^4.0.8"
   },
