+++ my-financial-vite/src/components/Home/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Home.css';
+
+function Home() {
+    return (
+        <div>
+            <p>HOME PAGE</p>
+        </div>
+    )
+}
+
+export default Home;
