+++ my-financial-vite/src/components/Investment/Investment.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Investment.css';
+
+function Investment() {
+    return (
+        <div>
+            <p>Investment PAGE</p>
+        </div>
+    )
+}
+
+export default Investment;
