+++ my-financial-vite/src/components/Login/Login.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Login.css';
+
+function Login() {
+    return (
+        <div>
+            <p>LOGIN PAGE</p>
+        </div>
+    )
+}
+
+export default Login;
