+++ my-financial-vite/src/components/Retirement/Retirement.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Retirement.css';
+
+function Retirement() {
+    return (
+        <div>
+            <p>RETIREMENT PAGE</p>
+        </div>
+    )
+}
+
+export default Retirement;
