+++ my-financial-vite/src/components/Signup/Signup.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Signup.css';
+
+function Signup() {
+    return (
+        <div>
+            <p>SIGN UP PAGE</p>
+        </div>
+    )
+}
+
+export default Signup;
